+++ old_packup/ExternalWrapper.hh
@@ -43,7 +43,7 @@ public:
     bool solve_leximax();
 
     virtual IntVector& get_model()          { return model; }
-    virtual void set_leximax_model(const std::vector<long long> &lib_model) {
+    virtual void set_leximax_model(const std::vector<int> &lib_model) {
         model.assign(lib_model.begin(), lib_model.end()); // this allows type conversion (if LINT is not long long)
     }
